Resolve dotnet path from DOTNET_HOST_PATH in VSTestTask
VSTestTask (legacy) was hardcoding 'dotnet' as the process filename,
assuming it would be found on PATH. This changes it to resolve the
dotnet executable using the same strategy as VSTestTask2:

1. Check DOTNET_HOST_PATH environment variable
2. Check current directory
3. Search PATH
4. Fall back to bare 'dotnet' name
