+++ src/Microsoft.TestPlatform.Build/Tasks/TestTaskUtils.cs
@@ -3,6 +3,8 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Runtime.InteropServices;
 
 using Microsoft.Build.Utilities;
 
@@ -236,5 +238,51 @@ internal static class TestTaskUtils
         }
 
         return builder.ToString();
+    }
+
+    /// <summary>
+    /// Resolves the full path to the dotnet executable by checking DOTNET_HOST_PATH,
+    /// current directory, and PATH in that order.
+    /// </summary>
+    /// <param name="dotnetExe">The platform-specific dotnet executable name (e.g. "dotnet.exe" or "dotnet").</param>
+    /// <returns>The resolved full path, or <see langword="null"/> if not found.</returns>
+    internal static string? ResolveDotnetPath(string dotnetExe)
+    {
+        //TODO: https://github.com/dotnet/sdk/issues/20 Need to get the dotnet path from MSBuild?
+
+        var dotnetHostPath = Environment.GetEnvironmentVariable("DOTNET_HOST_PATH");
+        if (!dotnetHostPath.IsNullOrEmpty())
+        {
+            var path = Path.Combine(Path.GetDirectoryName(Path.GetFullPath(dotnetHostPath))!, dotnetExe);
+            if (File.Exists(path))
+            {
+                return path;
+            }
+        }
+
+        if (File.Exists(dotnetExe))
+        {
+            return Path.GetFullPath(dotnetExe);
+        }
+
+        var values = Environment.GetEnvironmentVariable("PATH");
+        foreach (var p in values!.Split(Path.PathSeparator))
+        {
+            var fullPath = Path.Combine(p, dotnetExe);
+            if (File.Exists(fullPath))
+            {
+                return fullPath;
+            }
+        }
+
+        return null;
+    }
+
+    /// <inheritdoc cref="ResolveDotnetPath(string)"/>
+    internal static string? ResolveDotnetPath()
+    {
+        var dotnetExe = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? "dotnet.exe" : "dotnet";
+
+        return ResolveDotnetPath(dotnetExe);
     }
 }
